… Meta/Lagom/CMakeLists.txt
… Userland/Libraries/LibCrypto/BigInt/Algorithms/SimpleOperations.cpp
… Userland/Libraries/LibCrypto/BigInt/Algorithms/UnsignedBigIntegerAlgorithms.h
… Userland/Libraries/LibCrypto/CMakeLists.txt
… Userland/Libraries/LibCrypto/NumberTheory/ModularFunctions.cpp
… Userland/Utilities/test-crypto.cpp