# Overview

Followup to #2064 :^)

Adds an implementation the Montgomery Modular Power algorithm for a 5x-10x speedup of LibCrypto operations.

## Summary of changes

This pull request proposes :
- A reorganization of the non-allocating BigInt operations, by moving them out of the `UnsignedBigInteger` class and into a new `UnsignedBigIntegerAlgorithms` friend class and folder.
- A reorganization of some of the `NumberTheory` operations, by moving them into the `UnsignedBigIntegerAlgorithms` class.
- A new `add_into_accumulator_without_allocation` operation, that simplifies the code and removes the need for some of the temporary variables in other operations (this doesn't have a measurable performance impact).
- A new `montgomery_modular_power_with_minimal_allocations` operation, which allows for much faster computations in most asymmetric cryptographic operations.
- A change to the `STARTING_WORD_SIZE` constant which is now 32. This constant indicates how many stack-allocated words are given to BigIntegers by default. This new value was chosen to reduce stack impact in usual cases, but doesn't have any performance implications.


## Speedup overview

Here are the results of the speedup, obtained via Lagom by running `ctest -R LibCrypto` 10 times on my computer. Note that since the tests have high variance, we provide also the minimum and maximum time obtained for this run of tests.

|  Lagom Tests | Before the changes | After the changes | Speedup |
| ---- | ---- | ---- | ---- |
| Avg | 4.77s | 0.58s | 89% speedup (x8) |
| Max | 5.03s | 0.74s | 87% speedup (x6.5) |
| Min | 4.48s | 0.48s | 90% speedup (x9,3) |

In addition, here are the results obtained by running `test-crypto test` 10 times inside of Serenity.

|  Serenity Tests | Before the changes | After the changes | Speedup |
| ---- | ---- | ---- | ---- |
| Avg | 10.21s | 1.55s | 87% speedup (x6.5) |
| Max | 12.28s | 2.44s | 83% speedup (x5) |
| Min | 8.96s | 0.99s | 90% speedup (x9) |

When choosing the new value of the default stack allocations, we also ran several performance tests to see if there was an impact, and found none. Here are the results of these tests - the values given are from a series of 10 runs of `test-crypto test` directly inside of Serenity:

| value of STARTING_WORD_SIZE | 	Average | 	Maximum | 	Minimum |
| --- | ------ | ------ | ------ |
| 512 | 1.456s | 2.315s | 0.911s |
| 256 | 1.384s | 2.235s | 1.000s |
| 128 | 1.398s | 2.286s | 0.999s |
| 64  | 1.552s | 2.447s | 0.995s |
| 32  | 1.712s | 2.923s | 1.261s |
| 16  | 1.541s | 2.591s | 1.015s |
| 8   | 1.569s | 3.228s | 0.841s |
| 4   | 1.919s | 2.919s | 0.920s |
| 2   | 1.516s | 2.225s | 1.287s |
| 1   | 1.511s | 2.201s | 0.839s |

## Source of the algorithm

The Montgomery Modular Power algorithm was implemented based on :
 - [The monty_modpow implementation in the `num-bigint` Rust library](https://github.com/rust-num/num-bigint/blob/master/src/biguint/monty.rs) - this was the main source of inspiration and code samples to understand the algorithm's behavior, and the implementation fits the same overall structure as this file.
 - [This paper (PDF)](https://eprint.iacr.org/2011/239.pdf) describing about half of the algorithm both we and num-bigint use (Almost-Montgomery Multiplication).